// src/app/interfaces/shipping.interface.ts
export interface IShippingService {
  shipping_id: number;
  company_id: number;
  destination: string;
  min_weight: number;
  max_weight: number;
  usergroup_ids: string;
  rate_calculation: string;
  service_id: number;
  service_params: string;
  localization: string;
  tax_ids: string;
  position: number;
  status: string;
  free_shipping: string;
  is_address_required: string;

  created_at: string;
  updated_at: string;
}